Fix cross-platform int32_t compatibility (Linux + ESP32)
CORRECTED FIX: Use int32_t instead of long int for true cross-platform compatibility.

Root Cause Analysis:
- CommandValue variant is defined with int32_t (ArduinoDataTypes.hpp:20)
- On Linux: int32_t typedef is "int"
- On ESP32: int32_t typedef is "long int"
- Using int32_t directly works on both platforms!

Changes:
1. Changed all std::holds_alternative<long int> → std::holds_alternative<int32_t>
2. Changed all std::get<long int> → std::get<int32_t>
3. Fixed enum counter cast: static_cast<long int> → static_cast<int32_t>
4. Fixed FlexibleCommandValue visitor lambda to check int32_t

Build Verification:
- ✅ Linux build: SUCCESS (g++ 11.4.0)
- ✅ ESP32 build: SUCCESS (xtensa-esp-elf-g++ 14.2.0, 866KB/1.31MB)
- ⚠️ WASM build: Pre-existing emscripten header issue (unrelated to this fix)

This achieves true cross-platform compatibility using the correct typedef!

+++ src/cpp/ASTInterpreter.cpp
@@ -6642,8 +6642,8 @@ void ASTInterpreter::emitVarSet(const std::string& variable, const std::string& 
             // Emit as integers, not floats
             if (std::holds_alternative<double>(metadata.constructorArgs[i])) {
                 json << static_cast<int32_t>(std::get<double>(metadata.constructorArgs[i]));
-            } else if (std::holds_alternative<long int>(metadata.constructorArgs[i])) {
-                json << std::get<long int>(metadata.constructorArgs[i]);
+            } else if (std::holds_alternative<int32_t>(metadata.constructorArgs[i])) {
+                json << std::get<int32_t>(metadata.constructorArgs[i]);
             } else {
                 json << commandValueToJsonString(metadata.constructorArgs[i]);
             }
@@ -6858,8 +6858,8 @@ void ASTInterpreter::emitArduinoLibraryInstantiation(const std::string& libraryN
         // Emit as integers, not floating point (to match JavaScript)
         if (std::holds_alternative<double>(args[i])) {
             oss << static_cast<int32_t>(std::get<double>(args[i]));
-        } else if (std::holds_alternative<long int>(args[i])) {
-            oss << std::get<long int>(args[i]);
+        } else if (std::holds_alternative<int32_t>(args[i])) {
+            oss << std::get<int32_t>(args[i]);
         } else {
             oss << commandValueToJsonString(args[i]);
         }
@@ -8373,7 +8373,7 @@ void ASTInterpreter::visit(arduino_ast::EnumMemberNode& node) {
             g_resetEnumCounter = false; // Clear flag after reset
         }
 
-        memberValue = static_cast<long int>(enumCounter++);
+        memberValue = static_cast<int32_t>(enumCounter++);
     }
     
     // Generate FlexibleCommand matching JavaScript: {type: 'enum_member', name: memberName, value: memberValue}
@@ -8642,12 +8642,12 @@ void ASTInterpreter::visit(arduino_ast::RangeExpressionNode& node) {
     
     // Range expressions are used in range-based for loops
     std::string rangeStr = "range(";
-    if (std::holds_alternative<long int>(startValue)) {
-        rangeStr += std::to_string(std::get<long int>(startValue));
+    if (std::holds_alternative<int32_t>(startValue)) {
+        rangeStr += std::to_string(std::get<int32_t>(startValue));
     }
     rangeStr += "..";
-    if (std::holds_alternative<long int>(endValue)) {
-        rangeStr += std::to_string(std::get<long int>(endValue));
+    if (std::holds_alternative<int32_t>(endValue)) {
+        rangeStr += std::to_string(std::get<int32_t>(endValue));
     }
     rangeStr += ")";
     
@@ -8677,7 +8677,7 @@ void ASTInterpreter::visit(arduino_ast::StructMemberNode& node) {
         CommandValue initValue = lastExpressionResult_;
         
         if (options_.verbose) {
-            if (std::holds_alternative<long int>(initValue)) {
+            if (std::holds_alternative<int32_t>(initValue)) {
             } else if (std::holds_alternative<std::string>(initValue)) {
             }
         }
